# Conflicts:
#	README.md
#	api/controllers/hosts/edit.js
#	api/controllers/hosts/get.js
#	api/controllers/hosts/remove.js
#	api/helpers/recipients/add.js
#	api/helpers/recipients/remove.js
#	config/bootstrap.js

+++ api/controllers/directors/remove.js
@@ -15,7 +15,9 @@ module.exports = {
     },
 
     exits: {
-
+        conflict: {
+            statusCode: 409
+        }
     },
 
     fn: async function (inputs, exits) {
@@ -23,6 +25,15 @@ module.exports = {
         sails.log.silly(`Parameters passed: ${JSON.stringify(inputs)}`);
 
         try {
+            // First, determine if we need to lock out of editing admin
+            var lockout = await Directors.count({admin: true});
+
+            var toDestroy = await Directors.find({ID: inputs.ID});
+
+            // Block requests to remove this director if there are 1 or less admin directors and this director is an admin.
+            if (lockout <= 1 && toDestroy.admin)
+                return exits.conflict("To prevent accidental lockout, this request was denied because there are 1 or less admin directors. Make another director admin first before removing this director.");
+
             await Directors.destroy({ID: inputs.ID}).fetch();
 
             return exits.success();
